Refine SSH port handling and improve backup process reliability
- Cast `DB_BACKUP_SSH_PORT` to integer for stricter type safety.
- Update error handling in `isBackupHostReachable` method for clearer messages.
- Skip remote upload in local environments to streamline testing/debugging.

+++ config/backup-database.php
@@ -6,7 +6,7 @@ return [
 
     'ssh_host' => env('DB_BACKUP_SSH_HOST', '127.0.0.1'),
 
-    'ssh_port' => env('DB_BACKUP_SSH_PORT', 22),
+    'ssh_port' => (int) env('DB_BACKUP_SSH_PORT', 22),
 
     'ssh_user' => env('DB_BACKUP_SSH_USER'),
